Handle WebWorkers ourselves for Angular
