+++ src/alphaTab.main.ts
@@ -16,20 +16,10 @@ if (alphaTab.Environment.isRunningInWorker) {
                 );
             }
 
-            if (alphaTab.Environment.isWebPackBundled) {
-                alphaTab.Logger.debug('AlphaTab', 'Creating WebPack compatible worker');
-                // WebPack currently requires this exact syntax: new Worker(new URL(..., import.meta.url)))
-                // The module `@coderline/alphatab` will be resolved by WebPack to alphaTab consumed as library
-                // this will not work with CDNs because worker start scripts need to have the same origin like
-                // the current browser.
-
-                // https://github.com/webpack/webpack/discussions/14066
-
-                return new Worker(new URL('./alphaTab.worker', import.meta.url));
-            } else if (alphaTab.Environment.webPlatform == alphaTab.WebPlatform.BrowserModule) {
-                alphaTab.Logger.debug('AlphaTab', 'Creating Module worker');
-                return new Worker(new URL('./alphaTab.worker', import.meta.url), { type: 'module' });
-            }
+	        if (alphaTab.Environment.webPlatform == alphaTab.WebPlatform.BrowserModule || alphaTab.Environment.isWebPackBundled) {
+	            alphaTab.Logger.debug("AlphaTab", "Creating webworker");
+	            return new alphaTab.Environment.alphaTabWorker(new URL('./alphaTab.worker', import.meta.url), { type: 'module' });
+	        }
 
             // classical browser entry point
             if (!settings.core.scriptFile) {
@@ -58,14 +48,11 @@ if (alphaTab.Environment.isRunningInWorker) {
                 );
             }
 
-            if (
-                alphaTab.Environment.isWebPackBundled ||
-                alphaTab.Environment.webPlatform == alphaTab.WebPlatform.BrowserModule
-            ) {
-                alphaTab.Logger.debug('AlphaTab', 'Creating Module worklet');
-                const alphaTabWorklet = context.audioWorklet; // this name triggers the WebPack Plugin
-                return alphaTabWorklet.addModule(new URL('./alphaTab.worklet', import.meta.url));
-            }
+	        if (alphaTab.Environment.webPlatform == alphaTab.WebPlatform.BrowserModule || alphaTab.Environment.isWebPackBundled) {
+	            alphaTab.Logger.debug("AlphaTab", "Creating Module worklet");
+	            const alphaTabWorklet = context.audioWorklet; // this name triggers the WebPack Plugin
+	            return alphaTabWorklet.addModule(new URL('./alphaTab.worklet', import.meta.url));
+	        }
 
             alphaTab.Logger.debug('AlphaTab', 'Creating Script worklet');
             return context.audioWorklet.addModule(settings.core.scriptFile!);
