--- types/webpack/ModuleTypeConstants.d.ts
@@ -1,4 +1,0 @@
-declare module 'webpack/lib/ModuleTypeConstants' {
-    export const JAVASCRIPT_MODULE_TYPE_AUTO: string;
-    export const JAVASCRIPT_MODULE_TYPE_ESM: string;
-}
