--- types/webpack/index.d.ts
@@ -1,3 +1,0 @@
-import './ModuleTypeConstants'
-import './MakeSerializable'
-import './Identifier'
